ozon_ads_pipeline_full/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
pymongo>=4.8.0
tenacity>=8.5.0
icecream>=2.1.3
